+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] / 2026-05-14
+### Features
+- Support `Collectible` options to enable in the `AssemblyLoadContext` to allow unloading assemblies.
+### Updates
+- Update `ILTemplate` Collectible to `false`, disable `Unload` in `AssemblyLoadContext`.
+- Add `IsolatorExecute.Collectible` to update `Collectible` in the `ILTemplate`.
+
 ## [1.2.0] / 2026-04-29
 ### Features
 - Support build using `dotnet build`.
@@ -69,6 +76,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `Configuration` to use `LoadAtModuleInit` and force `Attach` to create default `AssemblyLoadContext` at module init.
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.3.0]: ../../compare/1.2.0...1.3.0
 [1.2.0]: ../../compare/1.1.0...1.2.0
 [1.1.0]: ../../compare/1.0.0...1.1.0
 [1.0.0]: ../../compare/1.0.0